Upload STaRT Specs Doc Final
This document details the technical specs of the STaRT application. 
Since there is no tracking for docx in GitHub, we decided to version control this document, put the corresponding versions in the branch MOT1. The final version after accepting all changes is put in Main. Version control details are provided on page 2 in this document.